small fixes and add styles
