+++ alerts/src/main/java/ru/touchin/roboswag/viewable_dialog/AlertDialogUtils.kt
@@ -1,11 +1,10 @@
-package ru.touchin.roboswag.alerts.dialog_view
+package ru.touchin.roboswag.viewable_dialog
 
-import android.app.AlertDialog
 import android.widget.TextView
 import androidx.core.view.isVisible
 import ru.touchin.extensions.setOnRippleClickListener
 
-fun setupButton(alertDialog: AlertDialog, buttonView: TextView, text: String?, onButtonClick: (() -> Unit)?) {
+fun setupButton(alertDialog: androidx.appcompat.app.AlertDialog, buttonView: TextView, text: String?, onButtonClick: (() -> Unit)?) {
     buttonView.setTextOrGone(text)
     buttonView.setOnRippleClickListener {
         onButtonClick?.invoke()
