+++ alerts/src/main/res/layout/dialog_alert.xml
@@ -10,13 +10,9 @@
 
     <TextView
         android:id="@+id/alert_title"
+        style="?attr/materialAlertDialogTitleTextStyle"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="24dp"
-        android:layout_marginRight="24dp"
-        android:paddingBottom="16dp"
-        android:textColor="@android:color/black"
-        android:textSize="18sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
@@ -24,14 +20,10 @@
 
     <TextView
         android:id="@+id/alert_message"
+        style="?attr/materialAlertDialogBodyTextStyle"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_below="@id/alert_title"
-        android:layout_marginLeft="24dp"
-        android:layout_marginRight="24dp"
-        android:lineSpacingExtra="8dp"
-        android:paddingBottom="28dp"
-        android:textSize="16sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/alert_title"
@@ -39,27 +31,23 @@
 
     <TextView
         android:id="@+id/alert_positive_button"
-        style="@style/DialogButton"
+        style="?attr/buttonBarPositiveButtonStyle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@id/alert_message"
         android:layout_alignParentEnd="true"
         android:layout_alignParentRight="true"
-        android:layout_marginTop="8dp"
-        android:layout_marginRight="8dp"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toBottomOf="@id/alert_message"
         tools:text="OK" />
 
     <TextView
         android:id="@+id/alert_negative_button"
-        style="@style/DialogButton"
+        style="?attr/buttonBarNegativeButtonStyle"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@id/alert_message"
-        android:layout_marginTop="8dp"
         android:layout_marginEnd="8dp"
-        android:layout_marginRight="8dp"
         android:layout_toStartOf="@id/alert_positive_button"
         android:layout_toLeftOf="@id/alert_positive_button"
         app:layout_constraintRight_toLeftOf="@id/alert_positive_button"
